# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RedSecTools API",
    "version": "2.0.0",
    "description": "Repository-only OpenAPI documentation for RedSecTools. This file is not served by the live application. It documents the current HTTP API, including auth requirements for main app, guest-link, extension, and admin routes. RedSecThreat now includes user-scoped keyword, tag, and alert ownership on top of shared feed ingestion, plus admin-managed feed, template, and notification policy routes. Survey and Wiki include their active create/manage/respond/results/documentation route groups alongside the existing Paste, Share, Vault, Chat, Calendar, Homepage, Extension, and Admin surfaces."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "UserSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "redsec_session",
        "description": "Signed user session cookie for the main web app."
      },
      "GuestSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "redsec_guest",
        "description": "Signed guest cookie created from /guest/{token}. Tool-scoped for paste or share."
      },
      "ExtensionBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque token",
        "description": "Bearer token returned by /api/ext/auth/login or /api/ext/auth/login/mfa."
      },
      "AdminCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "redsec_admin",
        "description": "Signed admin session cookie for /admin/api routes."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "retryAfter": { "type": "integer" },
          "restartLogin": { "type": "boolean" }
        },
        "required": ["error"]
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" }
        },
        "required": ["success"]
      },
...
      "ToolSummary": {
        "type": "object",
        "properties": {
          "key": { "type": "string" },
          "name": { "type": "string" },
          "href": { "type": "string" }
        },
        "required": ["key", "name", "href"]
      },
...
    },
    "responses": {
      "Success": {
        "description": "Successful operation",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Success" }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or expired",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not permitted",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/homepage/home-tab": {
      "get": {
        "tags": ["Homepage"],
        "summary": "Get homepage dashboard bootstrap data for the Home tab",
        "security": [{ "UserSessionCookie": [] }],
        "responses": {
          "200": {
            "description": "Home tab payload",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/homepage/bulletins": {
      "get": {
        "tags": ["Homepage"],
        "summary": "Get bulletin feed for the homepage dashboard",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "Bulletin feed payload",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Homepage"],
        "summary": "Create a bulletin",
        "security": [{ "UserSessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/homepage/bulletins/manage": {
      "get": {
        "tags": ["Homepage"],
        "summary": "Get bulletin management data for the current user",
        "security": [{ "UserSessionCookie": [] }],
        "responses": {
          "200": {
            "description": "Bulletin management payload",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/homepage/bulletins/{id}": {
      "get": {
        "tags": ["Homepage"],
        "summary": "Get a single bulletin",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Bulletin detail",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/BadRequest" }
        }
      },
      "put": {
        "tags": ["Homepage"],
        "summary": "Update a bulletin",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/BadRequest" }
        }
      },
      "delete": {
        "tags": ["Homepage"],
        "summary": "Delete a bulletin",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/homepage/bulletin-assets": {
      "post": {
        "tags": ["Homepage"],
        "summary": "Upload a bulletin image asset (converted to WebP)",
        "security": [{ "UserSessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": { "type": "string", "format": "binary" }
                },
                "required": ["image"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Uploaded bulletin asset",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "const": true },
                    "asset": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "url": { "type": "string" }
                      },
                      "required": ["id", "url"]
                    }
                  },
                  "required": ["success", "asset"]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/homepage/bulletin-assets/{id}": {
      "get": {
        "tags": ["Homepage"],
        "summary": "Fetch a bulletin asset image",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Bulletin asset image",
            "content": {
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/homepage/tool-favourites": {
      "get": {
        "tags": ["Homepage"],
        "summary": "Get the current user's built-in tool favourites",
        "security": [{ "UserSessionCookie": [] }],
        "responses": {
          "200": {
            "description": "Tool favourite selection",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "selected": {
                      "type": "array",
                      "items": { "type": "string" }
                    },
                    "available": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ToolSummary" }
                    }
                  },
                  "required": ["selected", "available"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "put": {
        "tags": ["Homepage"],
        "summary": "Update the current user's built-in tool favourites",
        "security": [{ "UserSessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "selected": {
                    "type": "array",
                    "maxItems": 4,
                    "items": { "type": "string" }
                  }
                },
                "required": ["selected"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Selection updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "const": true },
                    "selected": {
                      "type": "array",
                      "items": { "type": "string" }
                    }
                  },
                  "required": ["success", "selected"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  }
}